requirements.txt
tensorflow
tensorflow-datasets
numpy
matplotlib
scikit-learn
pillow
streamlit
opencv-python
pandas
